+++ src/class/panicMode/types.ts
@@ -1,0 +1,3 @@
+export type SetPanicModeToOptions = {
+    sendLogs: boolean;
+};