+++ test/helper/assertion.ts
@@ -1,3 +1,5 @@
+import { FlagshipSdkConfig } from '../../src/types';
+
 const assertionHelper = {
     getApiKeyHeader: (apiKey: string): { headers: { 'x-api-key': string } } => {
         return {
@@ -13,6 +15,9 @@ const assertionHelper = {
                 sendContextEvent: false
             }
         };
+    },
+    getTimeout: (url: string, config: FlagshipSdkConfig): { timeout: number } => {
+        return { timeout: url.includes('/campaigns') ? config.timeout * 1000 : undefined };
     }
 };